backend/requirements.txt
# NARUU AI Tourism Platform - Backend Dependencies

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10

# Validation & Serialization
pydantic==2.10.4
pydantic-settings==2.7.1
email-validator==2.2.0

# Authentication
PyJWT==2.10.1

# Environment
python-dotenv==1.0.1

# LINE Messaging API
line-bot-sdk==3.14.0

# AI
anthropic==0.44.0

# Translation
deepl==1.19.1

# HTTP Client
httpx==0.28.1

# Redis (optional)
redis[hiredis]==5.2.1

# Utilities
python-dateutil==2.9.0
orjson==3.10.14

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1

# Linting & Formatting
ruff==0.8.6
mypy==1.14.1
